+++ __tests__/stubs/artifact/internal/find/get-artifact.test.ts
@@ -1,9 +1,9 @@
 import { jest } from '@jest/globals'
 import * as core from '../../../../../__fixtures__/core.js'
-import { ResetCoreMetadata } from '../../../../../src/stubs/core.js'
+import { ResetCoreMetadata } from '../../../../../src/stubs/core/core.js'
 import { EnvMeta, ResetEnvMetadata } from '../../../../../src/stubs/env.js'
 
-jest.unstable_mockModule('../../../../../src/stubs/core.js', () => core)
+jest.unstable_mockModule('../../../../../src/stubs/core/core.js', () => core)
 
 const getArtifact = await import(
   '../../../../../src/stubs/artifact/internal/find/get-artifact.js'
