+++ __tests__/stubs/artifact/internal/shared/user-agent.test.ts
@@ -1,6 +1,6 @@
 import { jest } from '@jest/globals'
 import { getUserAgentString } from '../../../../../src/stubs/artifact/internal/shared/user-agent.js'
-import { ResetCoreMetadata } from '../../../../../src/stubs/core.js'
+import { ResetCoreMetadata } from '../../../../../src/stubs/core/core.js'
 import { ResetEnvMetadata } from '../../../../../src/stubs/env.js'
 
 describe('user-agent', () => {
