+++ __tests__/stubs/artifact/internal/upload/upload-artifact.test.ts
@@ -4,7 +4,7 @@ import * as core from '../../../../../__fixtures__/core.js'
 import * as crypto from '../../../../../__fixtures__/crypto.js'
 import * as fs from '../../../../../__fixtures__/fs.js'
 import * as stream from '../../../../../__fixtures__/stream/promises.js'
-import { ResetCoreMetadata } from '../../../../../src/stubs/core.js'
+import { ResetCoreMetadata } from '../../../../../src/stubs/core/core.js'
 import { EnvMeta, ResetEnvMetadata } from '../../../../../src/stubs/env.js'
 
 const validateArtifactName = jest.fn()
@@ -27,7 +27,7 @@ const writeStream = {
 jest.unstable_mockModule('crypto', () => crypto)
 jest.unstable_mockModule('fs', () => fs)
 jest.unstable_mockModule('stream/promises', () => stream)
-jest.unstable_mockModule('../../../../../src/stubs/core.js', () => core)
+jest.unstable_mockModule('../../../../../src/stubs/core/core.js', () => core)
 jest.unstable_mockModule(
   '../../../../../src/stubs/artifact/internal/upload/upload-zip-specification.js',
   () => ({
