+++ __tests__/stubs/env.test.ts
@@ -1,5 +1,5 @@
 import { jest } from '@jest/globals'
-import { ResetCoreMetadata } from '../../src/stubs/core.js'
+import { ResetCoreMetadata } from '../../src/stubs/core/core.js'
 import { EnvMeta, ResetEnvMetadata } from '../../src/stubs/env.js'
 import type { EnvMetadata } from '../../src/types.js'
