+++ src/commands/run.ts
@@ -2,7 +2,7 @@ import { config } from 'dotenv'
 import { createRequire } from 'module'
 import quibble from 'quibble'
 import { ARTIFACT_STUBS } from '../stubs/artifact/artifact.js'
-import { CORE_STUBS, CoreMeta } from '../stubs/core.js'
+import { CORE_STUBS, CoreMeta } from '../stubs/core/core.js'
 import { EnvMeta } from '../stubs/env.js'
 import type { Action } from '../types.js'
 import { printTitle } from '../utils/output.js'
