+++ src/stubs/artifact/internal/find/get-artifact.ts
@@ -4,7 +4,7 @@ import type { OctokitOptions } from '@octokit/core'
 import { requestLog } from '@octokit/plugin-request-log'
 import { retry } from '@octokit/plugin-retry'
 import { EnvMeta } from '../../../../stubs/env.js'
-import * as core from '../../../core.js'
+import * as core from '../../../core/core.js'
 import {
   ArtifactNotFoundError,
   InvalidResponseError
