+++ src/stubs/artifact/internal/find/list-artifacts.ts
@@ -4,7 +4,7 @@ import type { OctokitOptions } from '@octokit/core'
 import { requestLog } from '@octokit/plugin-request-log'
 import { retry } from '@octokit/plugin-retry'
 import { EnvMeta } from '../../../../stubs/env.js'
-import * as core from '../../../core.js'
+import * as core from '../../../core/core.js'
 import type { Artifact, ListArtifactsResponse } from '../shared/interfaces.js'
 import { getUserAgentString } from '../shared/user-agent.js'
 import { getRetryOptions } from './retry-options.js'
